requirements.txt
# Core dependencies
OpenBB>=4.0.0
pandas>=1.5.0
numpy>=1.23.0
ta>=0.10.2

# Data sources
yfinance>=0.2.28

# Communication
python-telegram-bot>=20.0
requests>=2.28.0

# Configuration & Environment
python-dotenv>=0.21.0
pyyaml>=6.0

# Web framework
flask>=2.3.0

# File watching (for config hot-reload)
watchdog>=3.0.0

# Task scheduling
apscheduler>=3.10.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Data visualization (Phase 6: Streamlit UI)
streamlit>=1.28.0
plotly>=5.14.0

# Performance
cython>=0.29.0
